[tests] Remove non-math dependencies from math tests
